Mid dev all socket tests pass
