+++ packages/server-bridge/src/factories/defaultStellarFactory.js
@@ -2,7 +2,7 @@ import { configureStellar } from '@stellarjs/core';
 import transportFactory from '@stellarjs/transport-bull';
 
 let stellarFactory = null;
-export default function connectToMicroservices(log) {
+export default function connectToMicroservices({ log }) {
   if (!stellarFactory) {
     // eslint-disable-next-line better-mutation/no-mutation
     stellarFactory = configureStellar({ log, transportFactory, sourcePrefix: `bridge-` });
