+++ packages/server-bridge/src/factories/getSessionFactory.js
@@ -1,0 +1,9 @@
+import get from 'lodash/get';
+
+import sessions from './sessions';
+
+export default function getSessionFactory() {
+  return function getSession(path) {
+    return get(sessions, path);
+  };
+}
