+++ packages/server-bridge/src/factories/getTxNameFactory.js
@@ -1,0 +1,9 @@
+export default function getTxNameFactory() {
+  return function getTxName(requestHeaders) {
+    if (requestHeaders.queueName) {
+      return `${requestHeaders.queueName}`;
+    }
+
+    return requestHeaders.type;
+  };
+}
