+++ packages/server-bridge/src/factories/sessions.js
@@ -1,0 +1,2 @@
+const sessions = {};
+export default sessions;
